+++ vm/srcs/commands/process_sti_run.c
@@ -20,13 +20,14 @@ void		process_sti_run(t_vm *vm, t_process *pr, t_runner *run)
 
 	val2 = get_reg_dir_ind_arg(pr, run, A2);
 	val3 = get_reg_dir_arg(pr, run, A3);
-	set_field_vals(vm->field, pr->pc + (val2 + val3) % IDX_MOD,
+	set_field_vals(vm->field, pr->pc + (int)(val2 + val3) % IDX_MOD,
 		pr, run->args[A1]);
 	if (vm->debug_args & VERB_SHOW_OPERATIONS)
 	{
 		ft_printf("P% 5d | sti r%ld %ld %ld\n", pr->id,
 			run->args[A1] + 1, val2, val3);
-		ft_printf("       | -> store to %ld + %ld = %ld (with pc and mod %ld)"
-			"\n", val2, val3, val2 + val3, pr->pc + (val2 + val3) % IDX_MOD);
+		ft_printf("       | -> store to %ld + %ld = %d (with pc and mod %d)"
+			"\n", val2, val3, (int)(val2 + val3),
+			pr->pc + (int)(val2 + val3) % IDX_MOD);
 	}
 }
